==> ro/archive2/3d-plotter-party.ipynb <==
import pandas as pd, numpy as np, json, os
import matplotlib.pyplot as plt
%matplotlib inline

plt.style.use('seaborn-whitegrid')

#!pip install d3IpyPlus
from d3IpyPlus import ScatterPlot, LinePlot, BarPlot, StackedArea

#!pip install pyecharts

m=json.loads(open('ignore/member_timelines.json','r').read())
members=json.loads(open('ignore/member_simple.json','r').read())

d=pd.DataFrame(members['deaths'],index=['v']).T

d['v']=' †'

s=pd.DataFrame(members['s2'],index=['s']).T
s=s.join(d)
s['v']=s['v'].fillna('')
s['s']=s['s']+s['v']
s=s[['s']]
s['w']=s['s'].str.split(' ').str[0].str[0]+s['s'].str.split(' ').str[-1].str[0]

szd_color='#E91C62'

pcolors=json.loads(open('json/pcolors.json','r').read())

g=pd.DataFrame(members['megye'],index=['Megye ']).T
g.head()

data={}
attention={}
for k,i in enumerate(m):
    if k%1000==0: print(k/len(m)*100,'%')
    if i['Típus']=='Párt':
        if i['Csoportok'] not in data:data[i['Csoportok']]={'start':pd.to_datetime('2030'),'end':pd.to_datetime('1980')}
        start=pd.to_datetime(i['start'])
        end=pd.to_datetime(i['end'])
        data[i['Csoportok']]['start']=min(data[i['Csoportok']]['start'],start)
        data[i['Csoportok']]['end']=max(data[i['Csoportok']]['end'],end)

part_timelines=pd.DataFrame(data).T
part_timelines.index.name='id'

part_timelines['start']=part_timelines['start'].astype(str).str[:10]
part_timelines['end']=part_timelines['end'].astype(str).str[:10]

p=part_timelines.reset_index()[['id']]
p['Color']=p['id'].str.split(' ').str[1].str.strip()
p['Part']=p['id'].str.split('-').str[0].str.strip()

part_timelines=part_timelines.join(p.set_index('id').drop_duplicates())

logos={'🕊️':'ALDE', '🌹':'PSD', '📐':'PNL', '🔶':'PD', '⚛️':'PP-DD', '👤':'Független', '⚪️':'FSN',
       '✳️':'PNTCD', '🔱':'PRM', '🌷':'RMDSZ', '🍀':'ECO', '🍏':'PMP',
       '🔷':'USR', '🦅':'UNPR', '🔘':'Más'}

part_timelines['logo']=part_timelines['Part'].str.split(' ').str[0]

part_timelines['group']=[logos[i] for i in part_timelines['logo'].values]
part_timelines['label']=[i+' '+logos[i] for i in part_timelines['logo'].values]

part_timelines=part_timelines.join(pd.DataFrame(pcolors,index=['cvalue']).T,on='group').fillna('grey')

part_timelines_list=list(part_timelines.reset_index().T.to_dict().values())

open('json/part_timelines.json','w').write(json.dumps(part_timelines_list))

Resolve for non-activity


from calendar import monthrange

data={}
for k,i in enumerate(m):
    if k%1000==0: print(k/len(m)*100,'%')
    if i['Típus']=='Párt':
        part=i['Csoportok']
        if part not in data:data[part]={}
        for t in pd.date_range(i['start'],i['end']):
            st=str(t)[:7]
            if st not in data[part]:data[part][st]=0
            data[part][st]+=1/monthrange(t.year,t.month)[1]

df=pd.DataFrame(data)

df=df.stack().reset_index()
df.columns=['x','id','y']

df=df.groupby(['id','x']).mean().rolling(7).mean().reset_index().dropna()

df2=df.join(part_timelines,on='id')

df2=df2[df2['y']>5]
values=list(df2.T.to_dict().values())
open('json/part_timelines2.json','w').write(json.dumps(values))
